requirements.txt
pillow==10.0.1
pytesseract==0.3.10